Replace PokemonTypeLibrary project with package
`The project reference to PokemonTypeLibrary has been removed from the PkmnCalcWPF.sln solution file and its build configurations in the Global section. Instead, a package reference to PokemonTypeLibrary version 1.0.2 has been added to the PkmnCalcWPF.csproj project file. This change means that the PokemonTypeLibrary project will no longer be built with the solution, and the PkmnCalcWPF project will now use the PokemonTypeLibrary package.`